Reformat API list for Github
This changes the formatting of the API list so that it is easy to read on Github.
It looked fine in plain text, but ended up in a big blob when Github formatted it.

Also, I opted for `X` instead of [X] so that github wouldn't format some of the bullets as checkboxes but not others. I considered using emoji, but then the plaintext doesn't read well.